Better casting.
Former-commit-id: 16318ddc91312e7259628ff113596c5b094b4a86

+++ libdevcore/FixedHash.h
@@ -151,7 +151,7 @@ public:
 	{
 		FixedHash ret;
 		for (auto& i: ret.m_data)
-			i = std::uniform_int_distribution<uint16_t>(0, 255)(_eng);
+			i = std::uniform_int_distribution<uint8_t>(0, 255)(_eng);
 		return ret;
 	}
